new map size 320x240, changed uin8_t to uint16_t

+++ src/data-sturctures/Point.h
@@ -6,9 +6,9 @@
 class Point
 {
     public:
-    uint8_t x;
+    uint16_t x;
     uint8_t y;
-    Point(uint8_t = 0, uint8_t = 0);
+    Point(uint16_t = 0, uint16_t = 0);
 };
 
 #endif
