+++ src/output/Matrix_map.h
@@ -8,8 +8,8 @@
 class Matrix_map
 {
     uint8_t *_map;
-    uint8_t number_of_rows_map;
-    uint8_t number_of_col_map;
+    uint16_t number_of_rows_map;
+    uint16_t number_of_col_map;
     uint32_t _colour;
     TFT_eSPI *_tft;
 
@@ -18,10 +18,10 @@ class Matrix_map
     public:
     Matrix_map(TFT_eSPI *tft);
     void draw_map();
-    void set_map(uint8_t matrix_map[],uint8_t row, uint8_t col, uint8_t r, uint8_t g, uint8_t b);
+    void set_map(uint8_t matrix_map[],uint16_t row, uint16_t col, uint8_t r, uint8_t g, uint8_t b);
     uint8_t* get_current_map();
-    uint8_t get_current_map_rows();
-    uint8_t get_current_map_cols();
+    uint16_t get_current_map_rows();
+    uint16_t get_current_map_cols();
 
 };
